add Group class

+++ src/main/java/com/kupreychik/controller/TeacherController.java
@@ -1,15 +1,19 @@
 package com.kupreychik.controller;
 
-import com.kupreychik.dto.request.StudentRequest;
 import com.kupreychik.dto.request.TeacherRequest;
+import com.kupreychik.dto.response.ErrorResponse;
 import com.kupreychik.service.JsonParseService;
 import com.kupreychik.service.TeacherService;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import lombok.SneakyThrows;
 
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 
+import static com.kupreychik.consts.RegexConsts.SIGNWEB_FORMAT;
 import static com.kupreychik.consts.WebConsts.*;
 
 public class TeacherController implements HttpHandler {
@@ -40,17 +44,8 @@ public class TeacherController implements HttpHandler {
                 if (url.equals(TEACHERS_PATH)) {
                     var response = teacherService.getTeachers();
                     responseAsString = jsonParseService.writeToJson(response);
-                } else {
-                    if (url.contains("surname")) {
-                        String paramSurname = url.substring(url.indexOf("surname=") + 8, url.lastIndexOf("&"));
-                        String paramName = url.substring(url.indexOf("&name=") + 6);
-                        // Если в конце прописывать числа, то на выходе получаем другое имя. Почему так?
-                        responseAsString = teacherService.getTeacherBySurname(paramSurname, paramName);
-                    } else {
-                        String id = url.replace(TEACHERS_PATH + "/", "");
-                        responseAsString = teacherService.getTeacherById(Long.parseLong(id));
-                    }
-                }
+                } else responseAsString = getError();
+
                 sendOk(exchange, responseAsString);
 
             }
@@ -58,20 +53,36 @@ public class TeacherController implements HttpHandler {
             case DELETE -> {
                 String url = String.valueOf(exchange.getRequestURI());
                 String id = url.replace(TEACHERS_PATH + "/", "");
-                var responseAsString = teacherService.delete(Long.parseLong(id));
+                String responseAsString;
+
+                if(id.matches(SIGNWEB_FORMAT)) {
+                    responseAsString = teacherService.delete(Long.parseLong(id));
+                } else responseAsString = getError();
+
                 sendOk(exchange, responseAsString);
             }
 
             case PUT -> {
-                var teacherRequest = jsonParseService.readObject(exchange.getRequestBody(), StudentRequest.class);
-                String url = String.valueOf(exchange.getRequestURI());
-                String id = url.replace(TEACHERS_PATH + "/", "");
-                var responseAsString = teacherService.change(Long.parseLong(id), (TeacherRequest) teacherRequest);
-                sendOk(exchange, responseAsString);
+                try (InputStreamReader inputStream = new InputStreamReader(exchange.getRequestBody());
+                     BufferedReader bufferedReader = new BufferedReader(inputStream)){
+                    int b;
+                    StringBuilder buf = new StringBuilder();
+                    while ((b = bufferedReader.read()) != -1) {
+                        buf.append((char) b);
+                    }
+                    String url = String.valueOf(exchange.getRequestURI());
+                    String id = url.replace(TEACHERS_PATH + "/", "");
+
+                    String responseAsString;
+
+                    if(id.matches(SIGNWEB_FORMAT)) {
+                        responseAsString = teacherService.change(Long.parseLong(id), buf.toString());
+                    } else responseAsString = getError();
+
+                    sendOk(exchange, responseAsString);
+                }
             }
-
         }
-
     }
 
     private static void sendOk(HttpExchange exchange, String response) throws IOException {
@@ -79,5 +90,10 @@ public class TeacherController implements HttpHandler {
         exchange.sendResponseHeaders(200, response.length());
         exchange.getResponseBody().write(response.getBytes());
         exchange.close();
+    }
+
+    @SneakyThrows
+    private String getError(){
+        return jsonParseService.writeToJson(new ErrorResponse("Wrong request"));
     }
 }
